requirements.txt
Django==3.0.8
django-simple-history==2.11.0
django-session-timeout==0.1.0
